Added 32-bit functionality
Users can now input a 32-bit pattern from 0 to 4,294,967,295.
It was realized that the 10-bit slow division algorithm's
repetition iterator can be increased to 31 to allow the
algorithm to be able to traverse 32 bits (31 - 0). The divisor
was also shifted to 30 because the divisor is a constant 2, or
10 in binary. After the divisor shift, 10 binary will now occupy
the 30th and 31th bits.

A known issue is that the input display is messed up because
the value gets automatically signed extended after
2,147,483,647. However, the binary output remains correct.

Signed-off-by: PaimonCodes <cryptsGuard@gmail.com>