Removed input message
Input message was removed because the input value is displayed as signed extended
and therefore does not display the original intended value. As of now, I don't know
of any instruction or method that will not sign extend a 32-bit pattern when loaded
into a register for RISCV32I. "lwu" instruction is available for RISCV, but only for 64-bit
systems.

Signed-off-by: PaimonCodes <cryptsGuard@gmail.com>